pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hot-makler-bot"
version = "0.1.0"
description = "TG-бот посуточной аренды жилья на Фукуоке"
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.13.1",
    "aiogram-calendar==0.6.0",
    "gspread==6.1.2",
    "google-auth==2.35.0",
    "sqlalchemy==2.0.36",
    "aiosqlite==0.20.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "apscheduler==3.10.4",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
    "sentry-sdk==2.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-mock==3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
